+++ src/backend/magnify/core/exceptions.py
@@ -1,3 +1,0 @@
-"""
-Specific exceptions for the core application
-"""