+++ src/backend/magnify/core/utils.py
@@ -1,3 +1,0 @@
-"""
-Utils that can be useful throughout Magnify's core app
-"""