+++ src/demo/frontend/src/App.tsx
@@ -1,0 +1,9 @@
+import { TestButton } from "@jitsi-magnify/frontend";
+
+export default function App() {
+  return (
+    <div>
+      <TestButton />
+    </div>
+  );
+}