+++ src/frontend/jest.config.ts
@@ -1,0 +1,14 @@
+import type { Config } from "@jest/types";
+
+const config: Config.InitialOptions = {
+  verbose: true,
+  transform: {
+    "^.+\\.tsx?$": `ts-jest`,
+  },
+  moduleNameMapper: {
+    "\\.(css|less|scss|sss|styl)$": "jest-css-modules",
+  },
+  testEnvironment: 'jsdom',
+};
+
+export default config;