+++ src/frontend/src/components/TestButton/TestButton.stories.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import TestButton from "./TestButton";
+
+export default {
+  title: "TestButton",
+  component: TestButton,
+};
+
+export const Default = () => <TestButton />;